Fix 'Default export names and file names should match' by renaming exported consts
