+++ src/components/AuthenticatingInfo.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import styles from './styles.module.css';
 
-const AuthenticatingInfoComponent = (): React.ReactElement => (
+const AuthenticatingInfo = (): React.ReactElement => (
   <div className={styles['content-element']}>Haetaan kirjautumistietoja...</div>
 );
 
-export default AuthenticatingInfoComponent;
+export default AuthenticatingInfo;
