+++ src/components/Login.tsx
@@ -5,10 +5,10 @@ import LogoutInfo from './LogoutInfo';
 import AuthenticatingInfo from './AuthenticatingInfo';
 import WithAuth from '../clients/WithAuth';
 
-const LoginComponent = (): React.ReactElement => (
+const Login = (): React.ReactElement => (
   <DemoWrapper title="Keycloak-kuuntelija">
     {WithAuth(LogoutInfo, LoginInfo, AuthenticatingInfo)}
   </DemoWrapper>
 );
 
-export default LoginComponent;
+export default Login;
